refactor: classify keyboard helper, query, dispatch layer
